requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<0.32.0
pdfplumber>=0.11.0,<0.12.0
python-dateutil>=2.9.0
pydantic>=2.7.0,<3.0.0
typer>=0.12.0,<0.13.0
rich>=13.7.0,<14.0.0
PyPDF2>=3.0.0,<4.0.0
python-multipart>=0.0.20
httpx>=0.28.0
